Flytter transformer-kall inn i service
