+++ src/main/kotlin/no/nav/personbruker/varseldb/api/varsel/varselApi.kt
@@ -7,7 +7,6 @@ import io.ktor.response.*
 import io.ktor.routing.*
 import no.nav.personbruker.varseldb.api.common.exception.ExceptionResponseHandler
 import org.slf4j.LoggerFactory
-import java.time.LocalDateTime
 
 val log = LoggerFactory.getLogger(VarselService::class.java)
 
@@ -16,8 +15,7 @@ fun Route.varselApi(service: VarselService) {
     post("/varsel") {
         try {
             val varselDto = call.receive<VarselDTO>()
-            val varsel = VarselTransformer.fromDTO(varselDto)
-            service.createVarsel(varsel)
+            service.createVarsel(varselDto)
             call.respond(HttpStatusCode.Created)
         } catch(exception: Exception) {
             val errorCode = ExceptionResponseHandler.logExceptionAndDecideErrorResponseCode(log, exception)
